+++ app/layout.tsx
@@ -1,39 +1,65 @@
-'use client';
-
 import { ChatCtxProvider } from '@/context/ChatCtx';
-import './globals.css';
+
 import { Source_Sans_3 } from '@next/font/google';
-import { createPagesBrowserClient } from '@supabase/auth-helpers-nextjs';
-import { SessionContextProvider, Session } from '@supabase/auth-helpers-react';
-import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import { useState } from 'react';
+
 import 'react-tooltip/dist/react-tooltip.css';
-import { Database } from '@/types/database.supabase';
-import ToastProvider from '@/lib/ToastProvider';
-import { QueryClient, QueryClientProvider } from 'react-query';
 
-const queryClient = new QueryClient();
+import RootLayoutComponent from '@/components/home/RootLayoutComponent';
 
-const sourceSans3 = Source_Sans_3({ subsets: ['latin'] });
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'FrostByte',
+  description: 'The Next Generation Open-Source Instant Messaging Application',
+  openGraph: {
+    url: 'https://www.frostbyteapp.com/',
+    siteName: 'FrostByte',
+    images: [
+      {
+        url: 'https://drive.google.com/uc?id=109htcEWnRyq1f3SN7Wp3h-CNns1zmkKo',
+        width: 800,
+        height: 600,
+      },
+    ],
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'FrostByte',
+    description:
+      'The Next Generation Open-Source Instant Messaging Application',
+    siteId: '958819679992127488',
+    creator: '@stanley_tsonev',
+    creatorId: '958819679992127488',
+    images: [
+      'https://drive.google.com/uc?id=109htcEWnRyq1f3SN7Wp3h-CNns1zmkKo',
+    ],
+  },
+};
 
 export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const [supabaseClient] = useState(() => createPagesBrowserClient<Database>());
-  return (
-    <html lang="en">
-      <SessionContextProvider supabaseClient={supabaseClient}>
-        <ChatCtxProvider>
-          <QueryClientProvider client={queryClient}>
-            <body className={sourceSans3.className}>
-              <ToastProvider>{children}</ToastProvider>
-            </body>
-          </QueryClientProvider>
-        </ChatCtxProvider>
-      </SessionContextProvider>
-    </html>
-  );
+  return <RootLayoutComponent>{children}</RootLayoutComponent>;
+}
+{
+  /* <title>FrostByte</title>
+      <meta content="width=device-width, initial-scale=1" name="viewport" />
+      <meta name="description" content="Generated by create next app" />
+      <link rel="icon" href="/favicon.ico" />
+      <meta property="og:title" content="FrostByte" />
+      <meta
+        property="og:description"
+        content="The Next Generation Open-Source Instant Messaging Application"
+      />
+      <meta
+        property="og:image"
+        content="https://drive.google.com/uc?id=109htcEWnRyq1f3SN7Wp3h-CNns1zmkKo"
+      />
+      <meta property="og:url" content="https://www.frostbyteapp.com/" />
+      <meta property="og:type" content="website" />
+      <meta property="og:site_name" content="FrostByte" /> */
 }
