--- app/login/head.tsx
@@ -1,22 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>FrostByte</title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <meta name="description" content="Generated by create next app" />
-      <link rel="icon" href="/favicon.ico" />
-      <meta property="og:title" content="FrostByte" />
-      <meta
-        property="og:description"
-        content="The Next Generation Open-Source Instant Messaging Application"
-      />
-      <meta
-        property="og:image"
-        content="https://drive.google.com/uc?id=109htcEWnRyq1f3SN7Wp3h-CNns1zmkKo"
-      />
-      <meta property="og:url" content="https://www.frostbyteapp.com/" />
-      <meta property="og:type" content="website" />
-      <meta property="og:site_name" content="FrostByte" />
-    </>
-  );
-}
